+++ src/Daedalus.jl
@@ -10,8 +10,12 @@ include("Ode.jl")
 include("Events.jl")
 include("Model.jl")
 include("Ensemble.jl")
+include("DaedalusOutput.jl")
 include("Outputs.jl")
+include("Costs.jl")
 
-export Npi, ParamEffect, ReactiveTrigger, TimeTrigger, get_values, get_time
+export Npi, ParamEffect, ReactiveTrigger, TimeTrigger, get_values, get_times,
+       DaedalusOutput, get_data, get_incidence, get_epidemic_summary, get_life_years_lost,
+       get_costs
 
 end